+++ app/scripts/env.ts
@@ -6,6 +6,7 @@ const body              = document.body;
 const pjaxContainer     = '.js-pjax-container';
 const pjaxWrapper       = '.js-pjax-wrapper';
 
+const scrollTrigger     = 100; // in pixels
 const easing            = {
                             ease: [0.4, 0.0, 0.2, 1],
                             in: [0.0, 0.0, 0.2, 1],
@@ -13,7 +14,7 @@ const easing            = {
                             sharp: [0.4, 0.0, 0.6, 1]
                         }
 
-export { APP_NAME, html, body, pjaxWrapper, pjaxContainer, isDebug, easing };
+export { APP_NAME, html, body, pjaxWrapper, pjaxContainer, isDebug, easing, scrollTrigger };
 
 export function setDebug(status:boolean){
     isDebug = status;